\name{bams-package}
\alias{bams-package}
\alias{bams}
\docType{package}
\title{Breakpoint annotation model smoothing}
\description{Code and data to compare several change-point detection models
on DNA copy number profiles.}
\details{
\tabular{ll}{Package: \tab bams\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2014.12.22\cr
License: \tab GPL-3\cr
Title: \tab Breakpoint annotation model smoothing\cr
Suggests: \tab GLAD, DNAcopy, xtable, cghFLasso, flsa, cghseg, grid, gada,
plyr, ggplot2 (>= 0.9.0), RColorBrewer, reshape2, lattice, proto,
changepoint\cr
Depends: \tab R (>= 2.10)\cr
Additional_repositories: \tab http://r-forge.r-project.org\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


