\name{dnacopy.smoothvec}
\alias{dnacopy.smoothvec}
\title{dnacopy smoothvec}
\description{Smooth a profile using DNAcopy.}
\usage{dnacopy.smoothvec(profile, var, vals, ...)}
\arguments{
  \item{profile}{A profile data.frame.}
  \item{var}{Smoothness variable.}
  \item{vals}{Smoothness values.}
  \item{\dots}{Other arguments, passed to segment.}
}

\value{Matrix of smoothed profiles: nparam x nprobes.}

\author{Toby Dylan Hocking}





