\name{gada.results}
\alias{gada.results}
\title{gada results}
\description{Recover a matrix of smoothed signals from the gada results.}
\usage{gada.results(pro, fit.list)}
\arguments{
  \item{pro}{Profile data.frame.}
  \item{fit.list}{List of gada results. Each gada result is a list, one element for
each chromosome.}
}



\author{Toby Dylan Hocking}





