\name{pick.best.index}
\alias{pick.best.index}
\title{pick best index}
\description{Minimizer for local models, described in article section 2.3
"Picking the optimal model"}
\usage{pick.best.index(err)}
\arguments{
  \item{err}{Vector of errors to minimize.}
}

\value{Integer index of the minimal error.}

\author{Toby Dylan Hocking}





