\name{runglad}
\alias{runglad}
\title{runglad}
\description{Run glad to smooth a profile.}
\usage{runglad(profile, ...)}
\arguments{
  \item{profile}{Profile data.frame.}
  \item{\dots}{Smoothing parameter for glad.}
}

\value{Smoothing matrix nparam x nprobes.}

\author{Toby Dylan Hocking}





