\name{several.breakpoints}
\alias{several.breakpoints}
\title{several breakpoints}
\description{Zero-one loss when breakpoint annotations are considered to mean
presence of 1 or more breakpoints.}
\usage{several.breakpoints(counts, anns)}
\arguments{
  \item{counts}{Counts of breakpoints of model.}
  \item{anns}{Annotations.}
}

\value{Matrix of 0 and 1.}

\author{Toby Dylan Hocking}





