\name{bams-package}
\alias{bams-package}
\docType{package}
\title{Breakpoint annotation model smoothing}
\description{Data and code to allow
reproduction of
"Learning smoothing models with
breakpoint annotations."}
\details{
\tabular{ll}{Package: \tab bams\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Breakpoint annotation model smoothing\cr
Suggests: \tab GLAD, DNAcopy, xtable, cghFLasso, flsa, cghseg,
plyr, ggplot2, RColorBrewer, reshape2, lattice\cr
Depends: \tab R (>= 2.10)\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


